Server: Fix compiler warning w/ Oracle Dev. Studio
(introduced by 0c8c01e2981cbb0d4a7bef35d12734824fd34a55)

+++ unix/Xvnc/programs/Xserver/hw/vnc/init.c
@@ -115,7 +115,7 @@ static Bool initOutputCalled = FALSE;
 static Bool noCursor = FALSE;
 char *desktopName = DEFAULT_DESKTOP_NAME;
 int traceLevel = 0;
-int rfbLEDState = rfbLEDUnknown;
+int rfbLEDState = (int)rfbLEDUnknown;
 
 char rfbThisHost[256];
